refactor(engine): 文件级决策移到读线程派发之前，消除工作线程串行段
原来的结构把"编码器选型 + QUAL 先验训练"这件文件级的事，寄生在"0 号
压缩线程处理第 0 块"这个事件上：id > 0 的工作线程一启动就 conditionVar.wait，
等 0 号线程跑完整块（判决 + 训练 + 压缩）才被 notify 放行。

由此派生的四个问题都是这个位置错误的必然推论，不是四个独立缺陷：

1. 判决器手上只有块 0，看不到文件总长，"先验是否划算"这类按全文件数据量
   做的判断无从立足。
2. 训练量被块大小绑架，同一份数据在不同 -l 档位得到不同质量的先验。
3. N-1 个线程等的不是判决（约 0.5s），是判决+训练+整块压缩（约 3.3s），
   实测每个工作线程 31% 的采样落在 conditionVar.wait 上。
4. "判决未定前不并发"这条不变量靠 notify 的时机成立，不靠数据流位置。
   另有一个隐患：若首块 preAnalysis 失败，预处理会改由某个并行工作线程在
   它自己那一块上触发，采样内容随调度漂移，编码器选择变成非确定行为。

改为在读线程把第一个数据块送进 inputDataPool 之前完成决策。先后关系由
数据流位置保证——工作线程只能从队列取块，取不到就无块可压；队列的入队/出队
自带 release/acquire，决策结果的可见性一并解决。于是 mutex/conditionVar/
syncFlag/isNeedNotify 以及 Actuator::getNotifyFlag 整套机制连同各执行器里的
notifyFlag 一并删除。同步辅助块发射时写线程尚未收到任何数据块，"辅助块物理上
先于全部数据块"从时序巧合变成位置事实。

顺带修掉两处同源的重复：

- readOneBlock 此前在 PbgzEngine/CodecEngine/IndexEngine 里有三份近乎重复的
  实现，后两份整体重写、不调基类，任何加在骨架上的改动都会漏掉两份（本次的
  决策钩子第一次加上去就没被执行）。现收敛成一份骨架加两个钩子：intakeBlock
  表达"这个块归不归我"，readBlockPreProc 捕捉只在读前有效的量。
- "辅助块不进数据流"这条规则对所有引擎都成立，此前 CodecEngine 用
  isAuxiliaryBlock 判了一遍、IndexEngine 又用写死的 == REFERENCE 判了半遍，
  于是新增的 QUAL_PRIOR 块直接让 index 命令报错中断。规则上提到读循环骨架，
  只留一处权威定义，index 命令随之恢复。

验证：l5 + 参考压 323MB SAM，产物与改动前逐字节相同（30,889,648），
往返 md5 e01e17397897c28b0ca9dae434a9de67；FASTQ 往返 md5 320b61649f04...；
index 生成正常；单元测试 298 通过（ReferenceMakeIndex 为既有失败）。
压缩 CPU 占用从 199% 升到 274%。

本次连带提交此前已完成但尚未落库的两批改动（文件层面与本次交织，无法拆分）：
QUAL 先验块的训练/发射/取用，以及 coder 层 check_exit 由 _Exit 改为抛
coder_exception、在工作线程收口。

+++ src/aux_block_consumer.h
@@ -61,7 +61,7 @@ public:
      * 且数据块 meta 要改为记录这个序号；块 meta 里的偏移只作 seek 手段与校验，
      * 不承担索引职责。
      */
-    virtual bool claim(RoughIOBlock* blockPtr, int32_t packageIndex) = 0;
+    virtual bool claim(RoughIOBlock* blockPtr, int64_t packageIndex) = 0;
 };
 
 /* 辅助块载荷按引用共享：解码中的数据块必须能让它引用的那份数据活到自己解完。 */
